apps/data-processing/requirements.txt
apache-airflow==2.8.1
pandas==2.0.3
boto3==1.26.137
pika==1.3.2
redis==4.5.5
SQLAlchemy==1.4.50
psycopg2-binary==2.9.6
requests==2.31.0
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.0
black==23.3.0
flake8==6.0.0
